requirements.txt
disnake>=2.10.1,<3.0
python-dotenv>=1.0.1,<2.0
